Pull more JSX out of helper files
