+++ editor/src/views/projects/projectView/ProjectComponents/ComponentOptionWithIcon.jsx
@@ -5,7 +5,7 @@ import { Box } from "@mui/material";
  * @param {Object} option - Autocomplete option object with label, value, and data about component type
  * @return {JSX.Element}
  */
-export const ComponentOptionWithIcon = ({ option, props }) => {
+const ComponentOptionWithIcon = ({ option, props }) => {
   const { data: { line_representation = null } = {} } = option;
 
   return (
@@ -34,3 +34,5 @@ export const ComponentOptionWithIcon = ({ option, props }) => {
     </Box>
   );
 };
+
+export default ComponentOptionWithIcon;
