+++ editor/src/views/projects/projectsListView/helpers.jsx
@@ -5,7 +5,8 @@ import Stack from "@mui/material/Stack";
 import Typography from "@mui/material/Typography";
 import CopyTextButton from "src/components/CopyTextButton";
 import parse from "html-react-parser";
-import { formatDateType, FormattedDateString } from "src/utils/dateAndTime";
+import { formatDateType } from "src/utils/dateAndTime";
+import FormattedDateString from "src/utils/FormattedDateString";
 import ProjectStatusBadge from "src/views/projects/projectView/ProjectStatusBadge";
 import RenderSignalLink from "src/components/RenderSignalLink";
 import { PROJECT_LIST_VIEW_QUERY_CONFIG } from "src/views/projects/projectsListView/ProjectsListViewQueryConf";
